Modified include statement for the run-time compiler (RTC) header file.
The file name was recently changed from FunctionRTC.hh to RTC_FunctionRTC.hh
within Trilinos.

I added the following in an attempt to include the proper file regardless
of which Trilinos version you're using:

This should work for most cases, but it probably will not work if you
are using a VOTD Trilinos (11.11.00) that is slightly out of date.  In
this case, you'll have Trilinos version 11.11.00, but the name change
for the file will not yet be in place for your version of the repo.  The
solution in that case is to update to the most current VOTD.

There were other, unrelated Trilinos build-system bugs that were causing
Peridigm build errors with very recent Trilinos VOTD builds.  So you
should update to the most current Trilinos VOTD anyway.


Former-commit-id: 7d1bd33469c08ca5be9bd60972c5aacfe6ffde1e