Remove unneeded library for Trilinos_LIBRARIES list
Explicitly removed PyTrilinos if it appears in Trilinos_LIBRARIES list.
Trilinos_LIBRARIES is automatically populated with all of the libraries that
were build when Trilinos was build, whether or not they are needed by Peridigm.
This usually does not cause any adverse issues, but in the case of PyTrilinos,
it links against Python.  Leaving it in the list causes Python to be a
depenedency of Peridigm which we don't want to force.


Former-commit-id: bfb159e84a4f1ae8c486e81f37af146e2a853220