C-style interface for invoking peridigm
Now Peridigm can be called from C code, using this function interface,
"run_peridigm(...)"

This is work in support of Hasan Jamal's (Purdue CS) task scheduling research.

The first time run_peridigm() gets called, the peridigm object is instantiated as
a static variable. For all subsequent calls to run_peridigm(), the
initialization steps are skipped. On the last call to run_peridigm()
the finalize flag must be set to true so that MPI_Finalize() gets called.

This is just the initial stub out of the interface. It will change as
the use cases evolve.


Former-commit-id: f9df905d0471b09014d27b233c34022f8c03fcba