Cleaning up recent commit on restart functionality.
Removed restart examples from examples directory.  This directory is intended
to be kept clean and provide real-world examples to users.  It is not used for
testing.

Added Contact_Perforation_With_Restart test.  The test is not currently enabled
because it does not pass.  The idea was to create a new version of the
Contact_Perforation test, and run it as two runs (with restart) instead of as a
single run.  The results from the restart version should match the results from
the single run.  Comparing results, this is not the case.  It appears that, at a minimum,
the velocity is not being properly initialized in the restart case.  The restart results
look plausible, but do not match up with the single-run case.

For the record, one can use the "algebra" seacas utility to cut the results file
from the full simulation into two parts for direct exodiff comparison with the results
from the restart runs.  Comments explaining how to do this are in the input deck
for the full simulation.

After looking over the restart commits up to this point (a.k.a. "Changing files"),
it looks like the state of restart is:

- Works only for serial.

- Not enabled for multi-physics.

- Operates by having one or more runs with start and end times that link up,
  for example run_1 running from time 0.0 to 1.2, and run_2 from 1.2 to 2.0.

- Does not play well with the output functionality tied to explicit dynamics.  This
  is really the fault of the output manager.  The issue is that the output does not
  appear to dump at the final time step.  Instead, it strictly follows the output
  frequency command provided by the user.  For example, if run_1 goes from time 0.0
  to 1.2, but the output frequency does not match up exactly such that output is
  requested at time 1.2, then there will be no output at the final time step (time
  1.2), which is what the subsequent restart expects, and thus there will be a
  mismatch and the restart will be initialized with data at slightly the wrong time
  step.  For the Contact_Perforation_With_Restart test, the output frequency was
  chosen such that the final step is output, which is a workaround for this issue.
  This workaround will work only for a user-defined time step.

- No error handling.

- Operates using matrix market files that are stored in a "restart" subdirectory.
  The simulation looks for files in this directory and, if present, uses them to
  initialize data.  Subsequent restarts write over the data in the "restart" directory.
  This results in the process being non-repeatable unless the "restart" directory is
  removed prior to running the first simulation.  Subsequent simulations must be
  run only once, and in the proper order.