Implement YAML input format capabilities -- closes #10
This implements an optional YAML input parser via the one developed for MueLu.
This creates an optional dependency on yaml-cpp that must be enabled during the
Trilinos configuration.  This will ultimately replace the *.peridigm input files
and currently Peridigm will print a "deprecation" message if a *.peridigm file
is used.  This will add more robustness, testing, and maintainability of the input
file parser without sacrificing too much additional markup form the existing text
file style.  XML format is not affected.

Squashed commit of the following:

commit e50a5bd637ee69826e36aad18b8462bc47b3b73b
Author: John T. Foster <johntfosterjr@gmail.com>
Date:   Mon Aug 15 17:17:25 2016 -0500

    Remove wrongly commited files in build

commit 6d8bd80792f80f79989d6703dc5ea59bc47ba80f
Author: John T. Foster <johntfosterjr@gmail.com>
Date:   Mon Aug 15 17:14:21 2016 -0500

    Add Yaml regression test

commit bfb8b40e8d52b11d9c3cfe06cf7e7cc87e40a49a
Author: John T. Foster <johntfosterjr@gmail.com>
Date:   Mon Aug 15 15:08:51 2016 -0500

    Add Yaml parser test

commit 714da11ecaa525a6f26deb0b0c22a711e6f6e816
Author: John T. Foster <johntfosterjr@gmail.com>
Date:   Mon Aug 15 15:01:18 2016 -0500

    yaml parser passing initial test

commit 9a506b79be7ead45ea798ec4f88fd8f385bcaf92
Author: John T. Foster <johntfosterjr@gmail.com>
Date:   Mon Aug 15 12:59:56 2016 -0500

    Add Muelu Yaml reader/parser