Adaptive Quasi-Static Load-Stepping Added.
Basically whenever a step is not converged, whether it's passed the
maximum number of iterations or the Belos solver fails to converge, the
load step is cut in half until convergence.

A verification test case is added including all the parameters
controlled by the user.

- Maximum load step reductions in one step
- Maximum total load step reductions
After any of these maximum allowable reductions is hit, then either the
run is terminated or a switch to the Verlet solver occurs (user
controls).

- Reduce all remaining load steps
If true, after one load-step is reduced, all the remaining steps are
reduced too.

- Adaptive output frequency
If true, after reducing the step size in half, the output frequency is
doubled up.

- Switch to Verlet
-- "Safety Factor" or "Fixed dt" to be set by the user. Otherwise,
a safety factor of 0.8 is used.
-- "Output Frequency" can be changed at the switch.