Replaced all .peridigm input files with .yaml input files in preparation for removing support for the .peridigm format (and eventual removal of the boost dependency).
Note that using the new .yaml input files requires that Trilinos be built with YAML support.

For users that have existing .peridigm input decks, a utility for converting them to .yaml is available at

scripts/peridigm_to_yaml.py

This script was used for conversion of all the .peridigm files in the repository, with the exception of
twist_and_pull.peridigm, which required manual editing to remove aprepro commands.  Aprepro is not supported
for .yaml or .xml input decks, and all aprepro support will go away when support for .peridigm files is removed.