Cleaned up cmake files in the test/verification directory.
Added individual CMakeLists.txt files for each test, which is the standard practice
and allows each test to be run individually with make test.

Corrected an error in the Contact_Friction_Time_Dependent_Coefficient test that was
causing it not to be run and re-blessed the gold file.

Activated the ep_cube test and re-blessed the gold file.

Deleted the Dakota_2x1x1 test.