Added damage model that computes damage at initialization using bond filters.
The behavior of this damage model is similar to putting a bond filter on the discretization.
In the case of applying a bond filter to the discretization, any bond that crosses a bond filter
is not created.  In the case of the damage model, all bonds are created, and those that cross a
bond filter are broken at the onset of the simulation.

The main motivation for this damage model is the application of a zero-flux boundary condition.