Bond-Associated Hypoelastic Correspondence Modeling is added.
A new model based on the following paper is added:
M. Behzadinasab and J.T. Foster (2020),
A semi-Lagrangian constitutive correspondence framework for peridynamics,
Journal of the Mechanics and Physics of Solids, 137:103862.

The bond-associated hypoelastic model is based on a non-local, bond-level
velocity gradient operator, which depends on the deformation of
neighbors of both its end-points. An average velocity gradient is defined
at node level for this purpose. At each time step, the node-level
velocity gradient is computed first, then synched between different
parallel processors (this piece was added to the main Peridigm.cpp file).

An elastic, elastic-plastic, and isotropic hardening correspondence
models are added for the new framework.

A new bond-associated damage correspondence modeling is also added to
the damage folder.

A verification test is added, which provides an input deck that is
suitable for the new framework.