Add stale actions
If a ticket is stale for 30 days, the bot will put a message there and close the ticket after an additional 5 days.
 
Pull request will get some stale message but will never be closed. 

The intention is to have fewer issues opened, since some are from 2016 and might not be valid with a newer version anymore.