Fixed a bug related to restart for explicit solver (#189)
There was a bug for the explicit solver that only shows up if restart
is invoked (the subsequent runs in restart). When a case is read from
restart files, the last step p is currently re-run. However since that
States N and NP1 are swapped before the restart files are written, this
would result in wrong calculations. A flag is added to bypass the force
calculation in the initialization step when restart files are read.