Bond-Associated Lagrangian Correspondence PD added (#201)
* Bond-Associated Lagrangian Correspondence PD added

A new material class is added based on the Lagrangian BA-PD.
Elastic, Elastic-Plastic, and Isotropic-Hardening materials are added
as a sub-class. The high-order implicit gradient functions are
incorporated to compute the deformation gradient, which is then used to
evaluate the Piola-Kirchhoff stress tensor. A new test is also included
to feature the input type for this new class of materials.

* A few fixes for the new model for improving code performance